REFACTOR: Renamed hooks and providers contents to use kebab case
