+++ __tests__/hooks/useDebounce.test.tsx
@@ -1,6 +1,6 @@
 import { act, renderHook } from "@testing-library/react";
 import { describe, expect, it, vi, afterEach } from "vitest";
-import useDebounce from "@/hooks/useDebounce";
+import useDebounce from "@/hooks/use-debounce";
 
 describe("useDebounce", () => {
   afterEach(() => {
