+++ __tests__/hooks/useOnPlay.test.tsx
@@ -1,6 +1,6 @@
 import { renderHook, act } from "@testing-library/react";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import useOnPlay from "@/hooks/useOnPlay";
+import useOnPlay from "@/hooks/use-on-play";
 import type { SongWithAlbum } from "../../types/music/song-with-album";
 import { createMockSongWithAlbum } from "../helpers/mockData";
