+++ app/albums/[id]/_components/album-detail-content.tsx
@@ -6,11 +6,11 @@ import { useRouter } from "next/navigation";
 import { Pencil, Trash2, MoreHorizontal } from "lucide-react";
 import { toast } from "sonner";
 import type { AlbumDetail } from "../../../../types/music/album-detail";
-import useLoadImage from "@/hooks/useLoadImage";
+import useLoadImage from "@/hooks/use-load-image";
 import { formatArtists } from "@/lib/music/format-artists";
 import { toSongsWithAlbum } from "@/lib/mappers/song";
 import SongsGrid from "@/components/song/songs-grid";
-import { useUser } from "@/hooks/useUser";
+import { useUser } from "@/hooks/use-user";
 import renameAlbum from "@/actions/album/rename-album";
 import deleteAlbum from "@/actions/album/delete-album";
 import { RenameAlbumSchema } from "@/schemas/albums/rename-album.schema";
