--- app/albums/components/AlbumsContent.tsx
@@ -1,29 +1,0 @@
-"use client";
-
-import type { AlbumWithArtists } from "../../../types/music/album-with-artists";
-import AlbumsGrid from "@/components/album/albums-grid";
-import AlbumsHeader from "./albums-header";
-
-interface AlbumsContentProps {
-  albums: AlbumWithArtists[];
-}
-
-/**
- * Main content component for the albums page.
- * Displays a header with an action button and a grid of albums.
- * 
- * @param albums - The list of albums with their associated artists.
- */
-const AlbumsContent: React.FC<AlbumsContentProps> = ({ albums }) => {
-  return (
-    <div className="px-6 pb-4">
-      <div className="flex items-center justify-between mt-4">
-        <h2 className="text-foreground text-xl font-semibold">Explore</h2>
-        <AlbumsHeader />
-      </div>
-      <AlbumsGrid albums={albums} />
-    </div>
-  );
-};
-
-export default AlbumsContent;
