+++ app/layout.tsx
@@ -1,9 +1,9 @@
 import Sidebar from "@/components/sidebar/sidebar";
 import "./globals.css";
 import { Figtree } from "next/font/google";
-import SupabaseProvider from "@/providers/SupabaseProvider";
-import UserProvider from "@/providers/UserProvider";
-import ModalProvider from "@/providers/ModalProvider";
+import SupabaseProvider from "@/providers/supabase-provider";
+import UserProvider from "@/providers/user-provider";
+import ModalProvider from "@/providers/modal-provider";
 import { Toaster } from "@/components/ui/sonner";
 import Player from "@/components/player/player";
 import { SidebarProvider, SidebarInset } from "@/components/ui/sidebar";
