+++ app/upload/page.tsx
@@ -5,8 +5,8 @@ import { useRouter } from "next/navigation";
 import { toast } from "sonner";
 import uniqid from "uniqid";
 import { ChevronLeft, Plus } from "lucide-react";
-import { useUser } from "@/hooks/useUser";
-import { useSessionContext } from "@/providers/SupabaseProvider";
+import { useUser } from "@/hooks/use-user";
+import { useSessionContext } from "@/providers/supabase-provider";
 import type { AlbumWithArtists } from "../../types/music/album-with-artists";
 import type { Artist } from "../../types/artist/artist";
 import { mapAlbumWithArtistsRow } from "@/lib/mappers/album";
