+++ components/mobile-nav/mobile-bottom-nav.tsx
@@ -29,9 +29,9 @@ import {
 import { Separator } from "@/components/ui/separator";
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { Button } from "@/components/ui/button";
-import { useUser } from "@/hooks/useUser";
-import useAuthModal from "@/hooks/useAuthModal";
-import { useSupabaseClient } from "@/providers/SupabaseProvider";
+import { useUser } from "@/hooks/use-user";
+import useAuthModal from "@/hooks/use-auth-modal";
+import { useSupabaseClient } from "@/providers/supabase-provider";
 import { toast } from "sonner";
 import { getInitials } from "@/lib/avatar/get-initials";
