+++ components/modals/create-album-modal.tsx
@@ -3,8 +3,8 @@
 import { useState, useEffect } from "react";
 import { toast } from "sonner";
 import uniqid from "uniqid";
-import { useSessionContext } from "@/providers/SupabaseProvider";
-import { useUser } from "@/hooks/useUser";
+import { useSessionContext } from "@/providers/supabase-provider";
+import { useUser } from "@/hooks/use-user";
 import type { AlbumWithArtists } from "../../types/music/album-with-artists";
 import type { Artist } from "../../types/artist/artist";
 import { ALBUM_WITH_ARTISTS_SELECT } from "@/actions/_db-selects";
