+++ components/player/player-layout-wrapper.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import usePlayer from "@/hooks/usePlayer";
+import usePlayer from "@/hooks/use-player";
 import { cn } from "@/lib/utils";
 
 /**
