+++ components/player/player.tsx
@@ -1,8 +1,8 @@
 "use client";
 
-import usePlayer from "@/hooks/usePlayer";
-import useGetSongById from "@/hooks/useGetSongById";
-import useLoadSongUrl from "@/hooks/useLoadSongUrl";
+import usePlayer from "@/hooks/use-player";
+import useGetSongById from "@/hooks/use-get-song-by-id";
+import useLoadSongUrl from "@/hooks/use-load-song-url";
 import PlayerContent from "./player-content";
 
 /**
