+++ components/player/playlist-panel.tsx
@@ -5,7 +5,7 @@ import { Check, Plus } from "lucide-react";
 import BounceLoader from "react-spinners/BounceLoader";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
-import useAddToPlaylist from "@/hooks/useAddToPlaylist";
+import useAddToPlaylist from "@/hooks/use-add-to-playlist";
 import PanelBackButton from "./panel-back-button";
 import type { Playlist } from "../../types/playlist/playlist";
