+++ components/song/songs-grid.tsx
@@ -4,7 +4,7 @@ import type { SongWithAlbum } from "../../types/music/song-with-album";
 import { cn } from "@/lib/utils";
 import { GRID_CLASSES } from "@/lib/grid-classes";
 import SongItem from "@/components/song/song-item";
-import useOnPlay from "@/hooks/useOnPlay";
+import useOnPlay from "@/hooks/use-on-play";
 
 interface SongsGridProps {
   songs: SongWithAlbum[];
