+++ hooks/use-get-song-by-id.ts
@@ -3,7 +3,7 @@
 import { useEffect, useMemo, useState } from "react";
 import { toast } from "sonner";
 import type { SongWithAlbum } from "../types/music/song-with-album";
-import { useSessionContext } from "@/providers/SupabaseProvider";
+import { useSessionContext } from "@/providers/supabase-provider";
 import { mapSongWithAlbumRow } from "@/lib/mappers/song";
 import { SONG_WITH_ALBUM_SELECT } from "@/actions/_db-selects";
