+++ providers/user-provider.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { MyUserContextProvider } from "@/hooks/useUser";
+import { MyUserContextProvider } from "@/hooks/use-user";
 
 interface UserProviderProps {
   children: React.ReactNode;
